feat: 헥스 색상 확장 추가 및 불필요 코드 제거
새로운 헥스 색상 초기화 extension을 별도 파일로 분리하여 재사용성을 개선함
ContentView에서 중복 및 사용되지 않는 테마 색상 설정 코드를 제거함
